# seng365_assignment1/app/resources/seng365_petition_site_api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SENG365 2020: Petition Site API",
    "version": "1.0.0",
    "description": "This specification documents the API that must be implemented for Assignment 1.\n\nThe API provides all of the server-side functionality necessary in order to implement the user stories for Assignment 2."
  },
  "servers": [
    {
      "url": "http://csse-s365.canterbury.ac.nz:4001/api/v1",
      "description": "reference server"
    }
  ],
...
  "paths": {
...
    "/users/{id}": {
      "parameters": [
        {
          "name": "id",
          "description": "ID of the user.",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/FullUser/properties/userId"
          }
        }
      ],
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Retrieve information about a user.",
        "description": "The email field is only included when the currently authenticated user is viewing their **own** details.",
        "security": [
          {},
          {
            "UserToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Change a user's details.",
        "description": "Only accessible by the user themselves. The email must be syntactically valid (e.g. it has to include an `@`, but whether or not it is actually a registered email address will not be checked). The email address must not already be in use. The password must not be empty. If the password is being changed, then the current password must also be provided.",
        "security": [
          {
            "UserToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "ChangeUserDetailsRequest",
                "properties": {
                  "name": {
                    "$ref": "#/components/schemas/User/properties/name"
                  },
                  "email": {
                    "$ref": "#/components/schemas/User/properties/email"
                  },
                  "password": {
                    "$ref": "#/components/schemas/FullUser/properties/password"
                  },
                  "currentPassword": {
                    "$ref": "#/components/schemas/FullUser/properties/password"
                  },
                  "city": {
                    "$ref": "#/components/schemas/User/properties/city"
                  },
                  "country": {
                    "$ref": "#/components/schemas/User/properties/country"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "title": "User",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Adam Anderson"
          },
          "city": {
            "type": "string",
            "minLength": 1,
            "example": "Christchurch",
            "nullable": true
          },
          "country": {
            "type": "string",
            "minLength": 1,
            "example": "New Zealand",
            "nullable": true
          },
          "email": {
            "type": "string",
            "minLength": 1,
            "format": "email",
            "example": "aaa11@uclive.ac.nz"
          }
        }
      },
...
    },
    "securitySchemes": {
      "UserToken": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      }
    }
  }
}